+++ etc/sql/sql.py
@@ -1,0 +1,30 @@
+
+
+class Petition():
+    shards = 5
+    number_of_replicas = '0-all'
+
+
+class Supporters():
+    shards = 5
+    number_of_replicas = '0-all'
+
+
+class Cities():
+    shards = 5
+    number_of_replicas = '0-all'
+
+
+class Users():
+    shards = 5
+    number_of_replicas = '0-all'
+
+
+class SSOTokens():
+    shards = 1
+    number_of_replicas = '0-all'
+
+
+class Lovely_ESSequences():
+    shards = 1
+    number_of_replicas = '0-all'