+++ iris/service/tests.py
@@ -19,6 +19,7 @@ def test_suite():
         create_crate_suite('petition/document.rst'),
         create_crate_suite('petition/service.rst'),
         create_crate_suite('petition/events.rst'),
+        create_crate_suite('petition/support.rst'),
         create_suite('petition/sm.rst'),
         create_crate_doc_suite('petitions.rst'),